+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.32] - 2023-08-23
+
 ### Fixed
 
 - Ensure no task can be executed when the pool is destroyed.